populating the track screen from saved entries

+++ src/com/ems/lifetracker/Metric.java
@@ -6,7 +6,7 @@ public class Metric {
 	private String desc;
 	private String unit; // name of unit (miles, hours, glasses, etc.)
 	private String type; //TODO: restrict this to binary/numeric/increment
-	private int dflt;
+	private double dflt;
 	//private String isBinaryType; // yes/no metrics, eg went out to lunch
 	//private String isNumericType; // enter a value metric, eg "9" for distance ran
 	//private String isIncrementType; // accumulated throughout metric, eg glasses of water
@@ -15,7 +15,7 @@ public class Metric {
 	
 	}
 	
-	public Metric(String name, String desc, String unit, String type, int dflt){
+	public Metric(String name, String desc, String unit, String type, double dflt){
 		this.name = name;
 		this.desc = desc;
 		this.unit = unit;
@@ -56,11 +56,11 @@ public class Metric {
 		return type;
 	}
 	
-	public void setDflt(int dflt) {
+	public void setDflt(double dflt) {
 		this.dflt = dflt;
 	}
 	
-	public int getDflt(){
+	public double getDflt(){
 		return dflt;
 	}
 }
